op-e2e/actions: remove the in-process op-geth L2 execution layer
The op-reth-test-engine subprocess is now the only EL backing L2Engine in the
action tests; the OP_E2E_ACTIONS_EL selector, the geth node/EngineApi backend,
the helpers/engineapi package, and the geth-only NewL2Engine jwtPath/EngineOption
parameters are gone from op-e2e/actions/helpers. The L1 side (L1Miner/L1Replica,
in-process geth as L1) is untouched.

The kona fault-proof action tests in rust/kona/tests/proofs still need an
in-process geth L2 engine: they read L2 blocks and state directly from the
*core.BlockChain to build execution witnesses, which the out-of-process reth
engine only exposes over RPC. That engine — the Engine API backend plus direct
BlockChain access — is relocated to op-e2e/e2eutils/gethengine, imported solely
by the kona proof harness, rather than deleted; op-e2e/actions no longer depends
on it.

Tests that existed to exercise the in-process op-geth engine from the action
tests go with it:
- RestartOpGeth (persistent-DataDir restart-recovery, meaningless for the
  ephemeral subprocess): the backend-agnostic restart-recovery behavior —
  op-node re-deriving a wiped engine from L1 — is covered by
  RestartEngineFreshState.
- enginetest/estimate_diff_test.go (byte-identical eth_estimateGas across the
  two backends): with one EL there is nothing to diff in-process. The estimator
  behavior itself stays pinned by the engine's unit tests and by every action
  test that embeds estimates into signed transactions.
- TestL2EngineAPI* stay: they assert engine-API semantics over RPC and now run
  against the reth engine.

op-e2e/actions is excluded from the go-tests sweeps (which have no engine
binary) and runs in the dedicated go-actions-tests-reth CI job instead.

The relocated geth EL is a stopgap: migrating the kona proof harness off the
in-process op-geth L2 EL onto the reth engine is tracked in #21905.
