+++ op-e2e/actions/derivation/reorg_test.go
@@ -3,7 +3,6 @@ package derivation
 import (
 	"math/big"
 	"math/rand"
-	"path"
 	"testing"
 
 	"github.com/ethereum-optimism/optimism/op-e2e/config"
@@ -13,9 +12,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/eth/ethconfig"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/node"
 	"github.com/stretchr/testify/require"
 
 	altda "github.com/ethereum-optimism/optimism/op-alt-da"
@@ -36,7 +33,6 @@ func TestReorgBatchType(t *testing.T) {
 		{"ReorgOrphanBlock", ReorgOrphanBlock},
 		{"ReorgFlipFlop", ReorgFlipFlop},
 		{"DeepReorg", DeepReorg},
-		{"RestartOpGeth", RestartOpGeth},
 		{"RestartEngineFreshState", RestartEngineFreshState},
 		{"ConflictingL2Blocks", ConflictingL2Blocks},
 		{"SyncAfterReorg", SyncAfterReorg},
@@ -575,108 +571,6 @@ type rpcWrapper struct {
 	client.RPC
 }
 
-// RestartOpGeth tests that the sequencer can restart its execution engine without rollup-node restart,
-// including recovering the finalized/safe state of L2 chain without reorging.
-func RestartOpGeth(gt *testing.T, deltaTimeOffset *hexutil.Uint64) {
-	if actionsHelpers.RethBackendSelected() {
-		// RestartOpGeth exercises op-geth's persistent-DataDir restart-and-recover: it closes the
-		// sequencer's execution engine and starts a fresh one on the same on-disk DB, asserting the
-		// finalized/safe L2 state is recovered from disk without a reorg. The out-of-process
-		// op-reth-test-engine is ephemeral — fresh state per process, no DataDir — so this
-		// op-geth-internal behavior cannot be reproduced. RestartEngineFreshState covers the
-		// backend-agnostic restart-recovery behavior (re-derivation from L1 onto a wiped engine)
-		// on both ELs; this test stays geth-gated until the in-process op-geth EL is removed.
-		gt.Skip("op-geth persistent-DB restart-recovery; not applicable to the ephemeral reth engine")
-	}
-	t := actionsHelpers.NewDefaultTesting(gt)
-	dbPath := path.Join(t.TempDir(), "testdb")
-	dbOption := func(_ *ethconfig.Config, nodeCfg *node.Config) error {
-		nodeCfg.DataDir = dbPath
-		return nil
-	}
-	dp := e2eutils.MakeDeployParams(t, actionsHelpers.DefaultRollupTestParams())
-	upgradesHelpers.ApplyDeltaTimeOffset(dp, deltaTimeOffset)
-	sd := e2eutils.Setup(t, dp, actionsHelpers.DefaultAlloc)
-	log := testlog.Logger(t, log.LevelDebug)
-	jwtPath := e2eutils.WriteDefaultJWT(t)
-	// L1
-	miner := actionsHelpers.NewL1Miner(t, log, sd.L1Cfg)
-	l1F, err := sources.NewL1Client(miner.RPCClient(), log, nil, sources.L1ClientDefaultConfig(sd.RollupCfg, false, sources.RPCKindStandard))
-	require.NoError(t, err)
-	// Sequencer
-	seqEng := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg, jwtPath, dbOption)
-	engRpc := &rpcWrapper{seqEng.RPCClient()}
-	l2Cl, err := sources.NewEngineClient(engRpc, log, nil, sources.EngineClientDefaultConfig(sd.RollupCfg))
-	require.NoError(t, err)
-	sequencer := actionsHelpers.NewL2Sequencer(t, log, l1F, miner.BlobStore(), altda.Disabled, l2Cl, sd.RollupCfg, sd.L1Cfg.Config, sd.DependencySet, 0)
-
-	batcher := actionsHelpers.NewL2Batcher(log, sd.RollupCfg, actionsHelpers.DefaultBatcherCfg(dp),
-		sequencer.RollupClient(), miner.EthClient(), seqEng.EthClient(), seqEng.EngineClient(t, sd.RollupCfg))
-
-	// start
-	sequencer.ActL2PipelineFull(t)
-
-	miner.ActEmptyBlock(t)
-
-	buildAndSubmit := func() {
-		// build some blocks
-		sequencer.ActL1HeadSignal(t)
-		sequencer.ActBuildToL1Head(t)
-		// submit the blocks, confirm on L1
-		batcher.ActSubmitAll(t)
-		miner.ActL1StartBlock(12)(t)
-		miner.ActL1IncludeTx(sd.RollupCfg.Genesis.SystemConfig.BatcherAddr)(t)
-		miner.ActL1EndBlock(t)
-		sequencer.ActL2PipelineFull(t)
-	}
-	buildAndSubmit()
-
-	// finalize the L1 data (first block, and the new block with batch)
-	miner.ActL1SafeNext(t)
-	miner.ActL1SafeNext(t)
-	miner.ActL1FinalizeNext(t)
-	miner.ActL1FinalizeNext(t)
-	sequencer.ActL1FinalizedSignal(t)
-	sequencer.ActL1SafeSignal(t)
-
-	// build and submit more
-	buildAndSubmit()
-	// but only mark the L1 block with this batch as safe
-	miner.ActL1SafeNext(t)
-	sequencer.ActL1SafeSignal(t)
-
-	// build some more, these stay unsafe
-	miner.ActEmptyBlock(t)
-	sequencer.ActL1HeadSignal(t)
-	sequencer.ActBuildToL1Head(t)
-
-	statusBeforeRestart := sequencer.SyncStatus()
-	// before restart scenario: we have a distinct finalized, safe, and unsafe part of the L2 chain
-	require.NotZero(t, statusBeforeRestart.FinalizedL2.L1Origin.Number)
-	require.Less(t, statusBeforeRestart.FinalizedL2.L1Origin.Number, statusBeforeRestart.SafeL2.L1Origin.Number)
-	require.Less(t, statusBeforeRestart.SafeL2.L1Origin.Number, statusBeforeRestart.UnsafeL2.L1Origin.Number)
-
-	// close the sequencer engine
-	require.NoError(t, seqEng.Close())
-	// and start a new one with same db path
-	seqEngNew := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg, jwtPath, dbOption)
-	// swap in the new rpc. This is as close as we can get to reconnecting to a new in-memory rpc connection
-	engRpc.RPC = seqEngNew.RPCClient()
-
-	// note: geth does not persist the safe block label, only the finalized block label
-	safe, err := l2Cl.L2BlockRefByLabel(t.Ctx(), eth.Safe)
-	require.NoError(t, err)
-	finalized, err := l2Cl.L2BlockRefByLabel(t.Ctx(), eth.Finalized)
-	require.NoError(t, err)
-	require.Equal(t, statusBeforeRestart.FinalizedL2, safe, "expecting to revert safe head to finalized head upon restart")
-	require.Equal(t, statusBeforeRestart.FinalizedL2, finalized, "expecting to keep same finalized head upon restart")
-
-	// sequencer runs pipeline, but now attached to the restarted geth node
-	sequencer.ActL2PipelineFull(t)
-	require.Equal(t, statusBeforeRestart.UnsafeL2, sequencer.L2Unsafe(), "expecting to keep same unsafe head upon restart")
-	require.Equal(t, statusBeforeRestart.SafeL2, sequencer.L2Safe(), "expecting the safe block to catch up to what it was before shutdown after syncing from L1, and not be stuck at the finalized block")
-}
-
 // RestartEngineFreshState tests that the rollup node recovers when its execution engine restarts
 // with all state lost (an ephemeral EL, or an operator wiping a datadir): the fresh engine boots at
 // genesis and the node re-derives the safe chain from L1, restoring safety and finality without a
@@ -687,13 +581,12 @@ func RestartEngineFreshState(gt *testing.T, deltaTimeOffset *hexutil.Uint64) {
 	upgradesHelpers.ApplyDeltaTimeOffset(dp, deltaTimeOffset)
 	sd := e2eutils.Setup(t, dp, actionsHelpers.DefaultAlloc)
 	log := testlog.Logger(t, log.LevelDebug)
-	jwtPath := e2eutils.WriteDefaultJWT(t)
 	// L1
 	miner := actionsHelpers.NewL1Miner(t, log, sd.L1Cfg)
 	l1F, err := sources.NewL1Client(miner.RPCClient(), log, nil, sources.L1ClientDefaultConfig(sd.RollupCfg, false, sources.RPCKindStandard))
 	require.NoError(t, err)
 	// Sequencer, with an engine that keeps no persistent state
-	seqEng := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg, jwtPath)
+	seqEng := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg)
 	engRpc := &rpcWrapper{seqEng.RPCClient()}
 	l2Cl, err := sources.NewEngineClient(engRpc, log, nil, sources.EngineClientDefaultConfig(sd.RollupCfg))
 	require.NoError(t, err)
@@ -747,7 +640,7 @@ func RestartEngineFreshState(gt *testing.T, deltaTimeOffset *hexutil.Uint64) {
 
 	// close the sequencer engine, and replace it with a fresh one: all execution-layer state is lost
 	require.NoError(t, seqEng.Close())
-	seqEngNew := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg, jwtPath)
+	seqEngNew := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg)
 	engRpc.RPC = seqEngNew.RPCClient()
 
 	// the fresh engine boots at genesis
@@ -778,8 +671,7 @@ func ConflictingL2Blocks(gt *testing.T, deltaTimeOffset *hexutil.Uint64) {
 	sd, _, miner, sequencer, seqEng, verifier, _, batcher := actionsHelpers.SetupReorgTestActors(t, dp, sd, log)
 
 	// Extra setup: a full alternative sequencer, sequencer engine, and batcher
-	jwtPath := e2eutils.WriteDefaultJWT(t)
-	altSeqEng := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg, jwtPath)
+	altSeqEng := actionsHelpers.NewL2Engine(t, log, sd.L2Cfg)
 	altSeqEngCl, err := sources.NewEngineClient(altSeqEng.RPCClient(), log, nil, sources.EngineClientDefaultConfig(sd.RollupCfg))
 	require.NoError(t, err)
 	l1F, err := sources.NewL1Client(miner.RPCClient(), log, nil, sources.L1ClientDefaultConfig(sd.RollupCfg, false, sources.RPCKindStandard))
