+++ op-e2e/actions/helpers/setups.go
@@ -15,7 +15,6 @@ import (
 )
 
 func SetupSequencerTest(t Testing, sd *e2eutils.SetupData, log log.Logger, opts ...SequencerOpt) (*L1Miner, *L2Engine, *L2Sequencer) {
-	jwtPath := e2eutils.WriteDefaultJWT(t)
 	cfg := DefaultSequencerConfig()
 	for _, opt := range opts {
 		opt(cfg)
@@ -25,7 +24,7 @@ func SetupSequencerTest(t Testing, sd *e2eutils.SetupData, log log.Logger, opts 
 
 	l1F, err := sources.NewL1Client(miner.RPCClient(), log, nil, sources.L1ClientDefaultConfig(sd.RollupCfg, false, sources.RPCKindStandard))
 	require.NoError(t, err)
-	engine := NewL2Engine(t, log.New("role", "sequencer-engine"), sd.L2Cfg, jwtPath, EngineWithP2P())
+	engine := NewL2Engine(t, log.New("role", "sequencer-engine"), sd.L2Cfg)
 	l2Cl, err := sources.NewEngineClient(engine.RPCClient(), log, nil, sources.EngineClientDefaultConfig(sd.RollupCfg))
 	require.NoError(t, err)
 
@@ -39,8 +38,7 @@ func SetupVerifier(t Testing, sd *e2eutils.SetupData, log log.Logger,
 	for _, opt := range opts {
 		opt(cfg)
 	}
-	jwtPath := e2eutils.WriteDefaultJWT(t)
-	engine := NewL2Engine(t, log.New("role", "verifier-engine"), sd.L2Cfg, jwtPath, EngineWithP2P())
+	engine := NewL2Engine(t, log.New("role", "verifier-engine"), sd.L2Cfg)
 	engCl := engine.EngineClient(t, sd.RollupCfg)
 	verifier := NewL2Verifier(t, log.New("role", "verifier"), l1F, blobSrc, altda.Disabled, engCl, sd.RollupCfg, sd.L1Cfg.Config, sd.DependencySet, syncCfg, cfg.SafeHeadListener)
 	return engine, verifier
