+++ rust/kona/tests/proofs/helpers/fixture.go
@@ -2,6 +2,7 @@ package helpers
 
 import (
 	"github.com/ethereum-optimism/optimism/op-e2e/actions/helpers"
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/gethengine"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/params"
@@ -15,7 +16,7 @@ type TestFixture struct {
 
 type FaultProofProgramL2Source struct {
 	Node        *helpers.L2Verifier
-	Engine      *helpers.L2Engine
+	Engine      *gethengine.L2Engine
 	ChainConfig *params.ChainConfig
 }
