+++ rust/kona/tests/proofs/holocene_helpers_test.go
@@ -2,6 +2,7 @@ package proofs
 
 import (
 	actionsHelpers "github.com/ethereum-optimism/optimism/op-e2e/actions/helpers"
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/gethengine"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-service/testlog"
 	"github.com/stretchr/testify/require"
@@ -20,7 +21,7 @@ type holoceneExpectations struct {
 	preHolocene, holocene expectations
 }
 
-func (h holoceneExpectations) RequireExpectedProgressAndLogs(t actionsHelpers.StatefulTesting, actualSafeHead eth.L2BlockRef, isHolocene bool, engine *actionsHelpers.L2Engine, logs *testlog.CapturingHandler) {
+func (h holoceneExpectations) RequireExpectedProgressAndLogs(t actionsHelpers.StatefulTesting, actualSafeHead eth.L2BlockRef, isHolocene bool, engine *gethengine.L2Engine, logs *testlog.CapturingHandler) {
 	t.Helper()
 
 	var exp expectations
